bus/heathzenith/h89: Remove signals present on a real h89bus
