+++ scripts/src/bus.lua
@@ -5767,8 +5767,6 @@ end
 
 if (BUSES["H89BUS"]~=null) then
 	files {
-		MAME_DIR .. "src/devices/bus/heathzenith/h89/cards.cpp",
-		MAME_DIR .. "src/devices/bus/heathzenith/h89/cards.h",
 		MAME_DIR .. "src/devices/bus/heathzenith/h89/cdr_fdc_880h.cpp",
 		MAME_DIR .. "src/devices/bus/heathzenith/h89/cdr_fdc_880h.h",
 		MAME_DIR .. "src/devices/bus/heathzenith/h89/h89bus.cpp",
@@ -5791,6 +5789,18 @@ if (BUSES["H89BUS"]~=null) then
 		MAME_DIR .. "src/devices/bus/heathzenith/h89/z_89_11.h",
 		MAME_DIR .. "src/devices/bus/heathzenith/h89/z37_fdc.cpp",
 		MAME_DIR .. "src/devices/bus/heathzenith/h89/z37_fdc.h",
+	}
+end
+
+---------------------------------------------------
+--
+--@src/devices/bus/heathzenith/h89/intr_cntrl.h,BUSES["HEATH_INTR_SOCKET"] = true
+---------------------------------------------------
+
+if (BUSES["HEATH_INTR_SOCKET"]~=null) then
+	files {
+		MAME_DIR .. "src/devices/bus/heathzenith/h89/intr_cntrl.cpp",
+		MAME_DIR .. "src/devices/bus/heathzenith/h89/intr_cntrl.h",
 	}
 end
