--- src/devices/bus/heathzenith/h89/cards.h
@@ -1,18 +1,0 @@
-// license:BSD-3-Clause
-// copyright-holders:R. Belmont
-/***************************************************************************
-
-    Heath/Zenith H-89 and Z-90 cards
-
-***************************************************************************/
-
-#ifndef MAME_BUS_HEATHZENITH_H89_CARDS_H
-#define MAME_BUS_HEATHZENITH_H89_CARDS_H
-
-#pragma once
-
-void h89_right_cards(device_slot_interface &device) ATTR_COLD;
-void h89_right_cards_mms(device_slot_interface &device) ATTR_COLD;
-void h89_right_p506_cards(device_slot_interface &device) ATTR_COLD;
-
-#endif // MAME_BUS_HEATHZENITH_H89_CARDS_H
